Adiciona e cria classes, adiciona readme atualizado, cria menu e cria função de adicionar usuários, projetos e atividades

+++ projeto/students/Doctors.java
@@ -1,7 +1,7 @@
 package projeto.students;
 
 public class Doctors extends Students {
-    public Doctors(int cpf, String type){
+    public Doctors(long cpf, String type){
         super(cpf, type);
     }
 }
