+++ projeto/students/Masters.java
@@ -2,7 +2,7 @@ package projeto.students;
 
 public class Masters extends Students{
     boolean isaMaster = true;
-    public Masters(int cpf, String type){
+    public Masters(long cpf, String type){
         super(cpf, type);
     }
